Add DailyPL's from Canvas
I just wanted to get them all in the same place.

This is a really rough conversion from HTML to gfm. For some reason pandoc was struggling to convert from an epub (which I could download from Canvas) so I ended up just copying the HTML directly from Canvas, cleaning it up a little, and using pandoc to convert it to gfm.

If I have time there is a lot more cleanup I should probably do.